backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
alembic==1.13.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
WTForms==3.1.1
pytest==7.4.3
pytest-flask==1.3.0
PyJWT==2.8.0
gunicorn==21.2.0 